Ex 2

+++ p3.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-	int i,j;
+	short i,j;
 	i = 1;
 	j = 2;
 	i++;
